api/requirements.txt
aiosmtplib==3.0.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.5.0
fpdf2==2.7.8
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-watch==4.2.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
websockets==15.0.1
pymysql==1.1.0
